requirements.txt
torch
torchvision
fastapi
uvicorn
opencv-python
pillow
